+++ NEWS.md
@@ -1,5 +1,9 @@
 # QFeatures 1.21
 
+## QFeatures 1.21.3
+
+- Add `readQFeaturesFromDIANN(multiplexing = 'dimethyl')` example.
+
 ## QFeatures 1.21.2
 
 - Add support for dimethyl multiplexing to `readQFeaturesFromDIANN()`